v0 without ssl

+++ ccb/src/main/java/ma/alakhdarbank/ccb/entity/Lot.java
@@ -5,6 +5,7 @@ package ma.alakhdarbank.ccb.entity;
 
 import java.util.Date;
 
+import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
@@ -26,7 +27,7 @@ public class Lot {
     @GeneratedValue
     private Long id;
 	
-	private Integer idLot;
+	private Long idLot;
 	
 	private Integer nbrCpt;
